refactor: extract GitHub data fetching into services/github.ts
Move the GitHub REST calls and response transformation out of
community/page.tsx and changelog/page.tsx into a dedicated service, leaving
both pages as pure consumers. Public surface is fetchCommunityData and
fetchChangelogEntries plus the data-shape interfaces; all helpers and raw API
types are module-private. Error handling and fetch revalidate options are
unchanged.

Addresses #1470
